assignment/assignment 1/requirements.txt
# Core dependencies
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.0.0

# Data handling
lxml==4.9.3
pycocotools==2.0.7
albumentations==1.3.1

# Training utilities
tensorboard==2.14.1
tqdm==4.66.1
pyyaml==6.0.1

# Evaluation
matplotlib==3.8.0
seaborn==0.13.0
scikit-learn==1.3.2

# Deployment
onnx==1.15.0
onnxruntime==1.16.3
onnxsim==0.4.35

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
